break up checks into individual files
